Reorder stuff, rename stuff, add license header
