+++ src/wave/wave-demuxer.ts
@@ -13,7 +13,7 @@ import { InputAudioTrack, InputAudioTrackBacking } from '../input-track';
 import { PacketRetrievalOptions } from '../media-sink';
 import { assert, UNDETERMINED_LANGUAGE } from '../misc';
 import { EncodedPacket, PLACEHOLDER_DATA } from '../packet';
-import { readAscii, readBytes, Reader, readU16, readU32, readU64 } from '../reader2';
+import { readAscii, readBytes, Reader, readU16, readU32, readU64 } from '../reader';
 
 export enum WaveFormat {
 	PCM = 0x0001,
